Fix to work with the newest versions of Sphinx and cornices.
Prior to this commit, cornice_sphinx was designed to work with
cornice < 2 and relied upon the API of the removed CorniceSchema
object. It also used a fake Env object instead of the Sphinx
BuildEnvironment object when rendering rst documents to Nodes.

This commit brings back some altered code that used to be part of
CorniceSchema so that we can continue to locate the child nodes. It
also deleted the fake Env object we'd been using in favor of using
the real BuildEnvironment object that Sphinx hands us. Finally, it
uses copies of the Node children rather than using them directly,
as Sphinx was failing to pickle the children objects.

fixes #10

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>